Move PathAndQueryMatcher to responseplayer_test.go
